move fsm to sepeare pkg
